Fix activating multiple sockets
Problem:

The underlying `libxpc_launch_activate_socket_trampoline` call returns
an array of C `int`s which are 4 bytes each (32 bit).  On 64 bit
architectures go's `int`s are 8 bytes each.

Solution:

Fix the type of `libxpc_launch_activate_socket()`s `c_fds` and switch
from `copy()`ing the whole array (since it doesn't actually match the
size of `fds`) to a `for` loop.  Since they're now different types
anyway, have it return `uintptr`s so they don't need to be cast later
on.

Background:

I believe the existing `libxpc_launch_activate_socket()` code mostly
works when a single socket is returned because it copies past the end of
the `c_fds` array when it fills the `fds` array and then later the
various go syscalls all cast back to `int32` which drop the garbage,
most significant 4 bytes (if the most significant garbage bit happend to
be 1 this would fail becuase the final `int32` would be negative, but in
practice the garbage bytes seem to be 0).

I noticed this when I deleted the `SockFamily` key from
"example/agent.plist".  This caused `launchd` to return 2 sockets.  I
was expecting `launchd.Activate()` to fail with `syscall.EINVAL` but
instead its internal `launchd.Sockets()` failed with:
> 2024/12/20 13:40:35 launchd.Socket failed: net.FileListener for 0
> failed: file file+net : getsockopt: socket operation on non-socket

This is because `fds` ended up being filled with: `[17179869187, 0]`.
That is: `fd[0] == 0x400000003` (which contains both the actual file
descriptors: 3 and 4) and `fd[1] == 0` (which is all garbage, but seems
to be 0s in practice).  `Sockets()` would actually successfully pass
that `fd[0]` through `net.FileListener()` (because it would truncate
down to fd 3), but then `fd[1]` would fail in `getsockopt()` because 0
is stdin's file descriptor and it's not a socket.

After these changes, running the example without `SockFamily` fails
(correctly) with:
> 2024/12/20 15:46:15 launchd.Socket failed: too many sockets: 2:
> invalid argument